backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
yt-dlp>=2024.11.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
faster-whisper>=1.0.0
edge-tts>=6.1.9
gtts>=2.5.0
Pillow>=10.0.0
python-multipart>=0.0.9
apscheduler>=3.10.0
openai>=1.0.0
anthropic>=0.18.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
